Move Eth Bridge events to the appropriate crate
