+++ crates/core/src/ethereum_structs.rs
@@ -7,163 +7,8 @@ use std::ops::{Add, AddAssign, Deref};
 
 use borsh::{BorshDeserialize, BorshSerialize};
 pub use ethbridge_structs::*;
-use namada_macros::BorshDeserializer;
-#[cfg(feature = "migrations")]
-use namada_migrations::*;
 use num256::Uint256;
 use serde::{Deserialize, Serialize};
-
-use crate::event::extend::EventAttributeEntry;
-use crate::event::{EventError, EventType};
-use crate::keccak::KeccakHash;
-
-/// Status of some Bridge pool transfer.
-#[derive(
-    Hash,
-    Clone,
-    Debug,
-    Eq,
-    PartialEq,
-    Ord,
-    PartialOrd,
-    BorshSerialize,
-    BorshDeserialize,
-    BorshDeserializer,
-    Serialize,
-    Deserialize,
-)]
-// TODO: move to `namada_ethereum_bridge::event` or
-// some similar path in the namada eth bridge crate
-pub enum BpTransferStatus {
-    /// The transfer has been relayed.
-    Relayed,
-    /// The transfer has expired.
-    Expired,
-}
-
-// TODO: move to `namada_ethereum_bridge::event` or
-// some similar path in the namada eth bridge crate
-impl From<BpTransferStatus> for EventType {
-    fn from(transfer_status: BpTransferStatus) -> Self {
-        (&transfer_status).into()
-    }
-}
-
-// TODO: move to `namada_ethereum_bridge::event` or
-// some similar path in the namada eth bridge crate
-impl From<&BpTransferStatus> for EventType {
-    fn from(transfer_status: &BpTransferStatus) -> Self {
-        match transfer_status {
-            BpTransferStatus::Relayed => event_types::BRIDGE_POOL_RELAYED,
-            BpTransferStatus::Expired => event_types::BRIDGE_POOL_EXPIRED,
-        }
-    }
-}
-
-// TODO: move to `namada_ethereum_bridge::event` or
-// some similar path in the namada eth bridge crate
-impl TryFrom<EventType> for BpTransferStatus {
-    type Error = EventError;
-
-    fn try_from(event_type: EventType) -> Result<Self, Self::Error> {
-        (&event_type).try_into()
-    }
-}
-
-// TODO: move to `namada_ethereum_bridge::event` or
-// some similar path in the namada eth bridge crate
-impl TryFrom<&EventType> for BpTransferStatus {
-    type Error = EventError;
-
-    fn try_from(event_type: &EventType) -> Result<Self, Self::Error> {
-        if *event_type == event_types::BRIDGE_POOL_RELAYED {
-            Ok(BpTransferStatus::Relayed)
-        } else if *event_type == event_types::BRIDGE_POOL_EXPIRED {
-            Ok(BpTransferStatus::Expired)
-        } else {
-            Err(EventError::InvalidEventType)
-        }
-    }
-}
-
-/// Ethereum bridge events on Namada's event log.
-#[derive(
-    Hash,
-    Clone,
-    Debug,
-    Eq,
-    PartialEq,
-    Ord,
-    PartialOrd,
-    BorshSerialize,
-    BorshDeserialize,
-    BorshDeserializer,
-    Serialize,
-    Deserialize,
-)]
-// TODO: move to `namada_ethereum_bridge::event` or
-// some similar path in the namada eth bridge crate
-pub enum EthBridgeEvent {
-    /// Bridge pool transfer status update event.
-    BridgePool {
-        /// Hash of the Bridge pool transfer.
-        tx_hash: KeccakHash,
-        /// Status of the Bridge pool transfer.
-        status: BpTransferStatus,
-    },
-}
-
-impl EthBridgeEvent {
-    /// Return a new Bridge pool expired transfer event.
-    pub const fn new_bridge_pool_expired(tx_hash: KeccakHash) -> Self {
-        Self::BridgePool {
-            tx_hash,
-            status: BpTransferStatus::Expired,
-        }
-    }
-
-    /// Return a new Bridge pool relayed transfer event.
-    pub const fn new_bridge_pool_relayed(tx_hash: KeccakHash) -> Self {
-        Self::BridgePool {
-            tx_hash,
-            status: BpTransferStatus::Relayed,
-        }
-    }
-}
-
-// TODO: move to `namada_ethereum_bridge::event::types` or
-// some similar path in the namada eth bridge crate
-pub mod event_types {
-    //! Ethereum bridge event types.
-
-    use super::EthBridgeEvent;
-    use crate::event::EventType;
-    use crate::event_type;
-
-    /// Bridge pool relay event.
-    pub const BRIDGE_POOL_RELAYED: EventType =
-        event_type!(EthBridgeEvent, "bridge-pool", "relayed");
-
-    /// Bridge pool expiration event.
-    pub const BRIDGE_POOL_EXPIRED: EventType =
-        event_type!(EthBridgeEvent, "bridge-pool", "expired");
-}
-
-// TODO: move to `namada_ethereum_bridge::event` or
-// some similar path in the namada eth bridge crate
-/// Extend an [`Event`](crate::event::Event) with Bridge pool tx hash data.
-pub struct BridgePoolTxHash<'tx>(pub &'tx KeccakHash);
-
-impl<'tx> EventAttributeEntry<'tx> for BridgePoolTxHash<'tx> {
-    type Value = &'tx KeccakHash;
-    type ValueOwned = KeccakHash;
-
-    const KEY: &'static str = "bridge_pool_tx_hash";
-
-    fn into_value(self) -> Self::Value {
-        self.0
-    }
-}
 
 /// This type must be able to represent any valid Ethereum block height. It must
 /// also be Borsh serializeable, so that it can be stored in blockchain storage.
